+++ CHANGELOG.md
@@ -39,6 +39,7 @@ The [history](HISTORY.md) contains the changelogs of older PostBOUND releases.
   The `MostCommonValues` can be used as a drop-in replacement for the old tuple-based API. In addition, it provides more
   high-level methods for working with the most common values.
 - Enabled the MySQL and DuckDB backends to fall back to emulated statistics if the database does not provide them.
+- The Postgres interface now has a `rollback()` method to put connections back into a valid state.
 - The Postgres statistics interface now consistently supports table references with a schema.
 - The `QueryPreparation` API now provides the `projection` and `output` parameters to modify the *SELECT* clause and the
   type of results to gather for all queries in a more flexible and intuitive way (how did *explain=True*  and